Update design and viewModel.
