+++ app/src/main/java/com/example/movie_application_esra_kaya/utils/Constants.kt
@@ -3,4 +3,5 @@ package com.example.movie_application_esra_kaya.utils
 object Constants {
     const val BASE_URL = "https://api.themoviedb.org/3/"
     const val API_KEY = "46d0b7df20b6dca0b123566e3d926b4b"
+    const val POSTER_PATH = "https://image.tmdb.org/t/p/w500/"
 }
